jsonrpc compiles
