+++ chain/warp_signature.go
@@ -1,0 +1,6 @@
+package chain
+
+type WarpSignature struct {
+	PublicKey []byte `json:"publicKey"`
+	Signature []byte `json:"signature"`
+}
