--- handlers/dependencies.go
@@ -1,20 +1,0 @@
-package listeners
-
-import (
-	"context"
-
-	"github.com/ava-labs/avalanchego/trace"
-	"github.com/ava-labs/avalanchego/utils/logging"
-	"github.com/ava-labs/hypersdk/chain"
-)
-
-type VM interface {
-	Tracer() trace.Tracer
-	Logger() logging.Logger
-	Registry() (chain.ActionRegistry, chain.AuthRegistry)
-	Submit(
-		ctx context.Context,
-		verifySig bool,
-		txs []*chain.Transaction,
-	) (errs []error)
-}
