+++ rpc/consts.go
@@ -1,0 +1,6 @@
+package rpc
+
+const (
+	JSONRPCEndpoint   = "/rpc"
+	WebSocketEndpoint = "/ws"
+)
