+++ vm/errors.go
@@ -8,10 +8,9 @@ import (
 )
 
 var (
-	ErrNotAdded       = errors.New("not added")
-	ErrDropped        = errors.New("dropped")
-	ErrNotReady       = errors.New("not ready")
-	ErrStateMissing   = errors.New("state missing")
-	ErrMessageMissing = errors.New("message missing")
-	ErrStateSyncing   = errors.New("state still syncing")
+	ErrNotAdded     = errors.New("not added")
+	ErrDropped      = errors.New("dropped")
+	ErrNotReady     = errors.New("not ready")
+	ErrStateMissing = errors.New("state missing")
+	ErrStateSyncing = errors.New("state still syncing")
 )
