+++ vm/resolutions.go
@@ -42,6 +42,14 @@ func (vm *VM) ChainID() ids.ID {
 	return vm.snowCtx.ChainID
 }
 
+func (vm *VM) NetworkID() uint32 {
+	return vm.snowCtx.NetworkID
+}
+
+func (vm *VM) SubnetID() ids.ID {
+	return vm.snowCtx.SubnetID
+}
+
 func (vm *VM) ValidatorState() validators.State {
 	return vm.snowCtx.ValidatorState
 }
@@ -264,6 +272,16 @@ func (vm *VM) IsValidator(ctx context.Context, nid ids.NodeID) (bool, error) {
 
 func (vm *VM) Proposers(ctx context.Context, diff int, depth int) (set.Set[ids.NodeID], error) {
 	return vm.proposerMonitor.Proposers(ctx, diff, depth)
+}
+
+func (vm *VM) CurrentValidators(
+	ctx context.Context,
+) (map[ids.NodeID]*validators.GetValidatorOutput, map[string]struct{}) {
+	return vm.proposerMonitor.Validators(ctx)
+}
+
+func (vm *VM) GatherSignatures(ctx context.Context, txID ids.ID, msg []byte) {
+	vm.warpManager.GatherSignatures(ctx, txID, msg)
 }
 
 func (vm *VM) NodeID() ids.NodeID {
